+++ hypothesis-python/docs/changelog.rst
@@ -18,6 +18,15 @@ Hypothesis 6.x
 
     .. include:: ../RELEASE.rst
 
+.. _v6.148.5:
+
+--------------------
+6.148.5 - 2025-12-01
+--------------------
+
+This patch improves the error message for :class:`~hypothesis.errors.FlakyStrategyDefinition`
+when the precondition for a rule is flaky (:issue:`4206`).
+
 .. _v6.148.4:
 
 --------------------